refactor(server): :recycle: Rename sendinblue connector to brevo and fix typing issues
