+++ apps/server/src/workflows/mail/addContact/addContact.ts
@@ -1,4 +1,4 @@
 import { AddContactRequest } from "@refugies-info/api-types";
-import { addToNewsletter } from "~/connectors/sendinblue/addToNewsletter";
+import { addToNewsletter } from "~/connectors/brevo/addToNewsletter";
 
 export const addContact = async (body: AddContactRequest): Promise<void> => addToNewsletter(body.email);
