+++ apps/server/src/workflows/mail/deleteContact/deleteContact.ts
@@ -1,4 +1,4 @@
-import { deleteFromNewsletterList } from "~/connectors/sendinblue/addToNewsletter";
+import { deleteFromNewsletterList } from "~/connectors/brevo/addToNewsletter";
 import { User } from "~/typegoose";
 
 export const deleteContact = async (user: User): Promise<void> => deleteFromNewsletterList(user.email);
