+++ apps/server/src/workflows/mail/isInContact/isInContact.ts
@@ -1,4 +1,4 @@
-import { isInNewsletterList } from "~/connectors/sendinblue/addToNewsletter";
+import { isInNewsletterList } from "~/connectors/brevo/addToNewsletter";
 import { User } from "~/typegoose";
 
 export const isInContact = async (user: User): Promise<boolean> => isInNewsletterList(user.email);
